0.2.0-GM
Cleanup old project files
