+++ CHANGELOG.md
@@ -1,7 +1,13 @@
-# 0.1.99 (Release Candidate 1)
-- Bring in core functionality
-
-# 0.1.99-demo-3
+# 0.2.0
+- Interface 2.0 basic layout.
+- Tooltips added.
+- Setting validation (NF)
+- Functionally unusable, demo is for layout to use set MainWindow2.xaml as entrypoint.
+- Tweak button icons by moving to WPF.UI icons vs converted SVGs. This resulted in much cleaner XAML
+- Add Info button which will eventually show version information and more.
+- Add LabelValidation() function to manage labels with ease using Dispatching.
+- Port DebugConsole() and ConsoleWriteLine() functions to new interface.
+- Version identification, TODO: Add version info to debug logs.
 - Additional info at top of log (include version for support purposes)
 - Set and import settings in settings page.
 - Validation functionality for all validation labels
@@ -15,19 +21,9 @@
 - Debug Log can be copied to clipboard with new "animated" copy to clipboard button.
 - Fixed log file naming to use small year, and 24hour time.
 - Debug field now readonly again.
-
-# 0.1.99-demo-2
-- Tweak button icons by moving to WPF.UI icons vs converted SVGs. This resulted in much cleaner XAML
-- Add Info button which will eventually show version information and more.
-- Add LabelValidation() function to manage labels with ease using Dispatching.
-- Port DebugConsole() and ConsoleWriteLine() functions to new interface.
-- Version identification, TODO: Add version info to debug logs.
-
-# 0.1.99-demo
-- Interface 2.0 basic layout.
-- Tooltips added.
-- Setting validation (NF)
-- Functionally unusable, demo is for layout to use set MainWindow2.xaml as entrypoint.
+- Remove artifacts from previous interface
+- Some code documentation
+- Bring in core functionality
 
 # 0.1.2 (hotfix-2)
 - [Hotfix]Fix a hang with MSFSLayoutGenerator.exe
