+++ VERSION.txt
@@ -1,1 +1,1 @@
-0.2.0-rc1
+0.2.0
